+++ packages/biometric-ed25519/src/fido2.ts
@@ -11,10 +11,10 @@ export class Fido2 {
             rpName,
             challengeSize: 128,
             attestation: 'none',
-            cryptoParams: [-8, -7],
+            cryptoParams: [-8, -7, -257],
             authenticatorAttachment: 'platform',
             authenticatorRequireResidentKey: true,
-            authenticatorUserVerification: 'discouraged'
+            authenticatorUserVerification: 'preferred'
         });
     }