+++ packages/near-api-js/src/connect.ts
@@ -1,12 +1,12 @@
 /**
  * Connect to NEAR using the provided configuration.
  *
- * {@link ConnectConfig.networkId} and {@link ConnectConfig.nodeUrl} are required.
+ * {@link ConnectConfig#networkId} and {@link ConnectConfig#nodeUrl} are required.
  *
  * To sign transactions you can also pass:
- * 1. {@link ConnectConfig.keyStore}
- * 2. {@link ConnectConfig.keyPath}
- * 3. {@link ConnectConfig.deps.keyStore} (deprecated, only for use in legacy applications)
+ * 1. {@link ConnectConfig#keyStore}
+ * 2. {@link ConnectConfig#keyPath}
+ * 3. {@link ConnectConfig#deps.keyStore} (deprecated, only for use in legacy applications)
  *
  * If all three are passed they are prioritize in that order.
  *
@@ -20,27 +20,60 @@
  *   })
  * }
  * ```
+ * @example disable library logs
+ * ```js
+ * async function initNear() {
+ *   const near = await connect({
+ *      networkId: 'testnet',
+ *      nodeUrl: 'https://rpc.testnet.near.org',
+ *      logger: false
+ *   })
+ * }
  * @module connect
  */
 import { readKeyFile } from './key_stores/unencrypted_file_system_keystore';
 import { InMemoryKeyStore, MergeKeyStore } from './key_stores';
 import { Near, NearConfig } from './near';
-import fetch from './utils/setup-node-fetch';
-import { logWarning } from './utils';
-
-global.fetch = fetch;
+import { Logger } from '@near-js/utils';
 
 export interface ConnectConfig extends NearConfig {
     /**
-     * Initialize an {@link key_stores/in_memory_key_store!InMemoryKeyStore} by reading the file at keyPath.
+     * Initialize an {@link InMemoryKeyStore} by reading the file at keyPath.
      */
     keyPath?: string;
 }
 
 /**
  * Initialize connection to Near network.
+ * @param config The configuration object for connecting to NEAR Protocol.
+ * @returns A Promise that resolves to a `Near` object representing the connection.
+ *
+ * @example
+ * ```js
+ * const connectionConfig = {
+ *   networkId: 'testnet',
+ *   nodeUrl: 'https://rpc.testnet.near.org',
+ *   walletUrl: 'https://wallet.testnet.near.org',
+ *   helperUrl: 'https://helper.testnet.near.org',
+ *   keyStore: new InMemoryKeyStore(),
+ *   deps: { keyStore: new BrowserLocalStorageKeyStore() },
+ *   logger: true,
+ *   keyPath: '/path/to/account-key.json',
+ *   masterAccount: 'master-account.near',
+ * };
+ *
+ * const nearConnection = await connect(connectionConfig);
+ * console.log(nearConnection); // Near object representing the connection
+ * ```
  */
 export async function connect(config: ConnectConfig): Promise<Near> {
+    if (config.logger === false) {
+        // disables logging
+        Logger.overrideLogger(undefined);
+    } else if (config.logger !== undefined && config.logger !== null) {
+        Logger.overrideLogger(config.logger);
+    }
+
     // Try to find extra key in `KeyPath` if provided.
     if (config.keyPath && (config.keyStore ||  config.deps?.keyStore)) {
         try {
@@ -57,12 +90,10 @@ export async function connect(config: ConnectConfig): Promise<Near> {
                     keyPathStore,
                     config.keyStore || config.deps?.keyStore
                 ], { writeKeyStoreIndex: 1 });
-                if (!process.env['NEAR_NO_LOGS']) {
-                    console.log(`Loaded master account ${accountKeyFile[0]} key from ${config.keyPath} with public key = ${keyPair.getPublicKey()}`);
-                }
+                Logger.log(`Loaded master account ${accountKeyFile[0]} key from ${config.keyPath} with public key = ${keyPair.getPublicKey()}`);
             }
         } catch (error) {
-            logWarning(`Failed to load master account key from ${config.keyPath}: ${error}`);
+            Logger.warn(`Failed to load master account key from ${config.keyPath}: ${error}`);
         }
     }
     return new Near(config);