+++ packages/transactions/src/actions.ts
@@ -19,6 +19,7 @@ export class AccessKeyPermission extends Enum {
 }
 
 export class AccessKey extends Assignable {
+    nonce: BN;
     permission: AccessKeyPermission;
 }