+++ packages/transactions/src/sign.ts
@@ -1,6 +1,6 @@
 import { Signer } from '@near-js/signers';
-import sha256 from 'js-sha256';
 import BN from 'bn.js';
+import { sha256 } from '@noble/hashes/sha256';
 
 import { Action, SignedDelegate } from './actions';
 import { createTransaction } from './create_transaction';
@@ -31,7 +31,7 @@ export interface SignedDelegateWithHash {
  */
 async function signTransactionObject(transaction: Transaction, signer: Signer, accountId?: string, networkId?: string): Promise<[Uint8Array, SignedTransaction]> {
     const message = encodeTransaction(transaction);
-    const hash = new Uint8Array(sha256.sha256.array(message));
+    const hash = new Uint8Array(sha256(message));
     const signature = await signer.signMessage(message, accountId, networkId);
     const signedTx = new SignedTransaction({
         transaction,
@@ -56,8 +56,9 @@ export async function signTransaction(...args): Promise<[Uint8Array, SignedTrans
 
 /**
  * Sign a delegate action
- * @params.delegateAction Delegate action to be signed by the meta transaction sender
- * @params.signer Signer instance for the meta transaction sender
+ * @options SignDelegate options
+ * @param options.delegateAction Delegate action to be signed by the meta transaction sender
+ * @param options.signer Signer instance for the meta transaction sender
  */
 export async function signDelegateAction({ delegateAction, signer }: SignDelegateOptions): Promise<SignedDelegateWithHash> {
     const message = encodeDelegateAction(delegateAction);
@@ -72,7 +73,7 @@ export async function signDelegateAction({ delegateAction, signer }: SignDelegat
     });
 
     return {
-        hash: new Uint8Array(sha256.sha256.array(message)),
+        hash: new Uint8Array(sha256(message)),
         signedDelegateAction,
     };
 }