+++ packages/utils/src/format.ts
@@ -1,4 +1,5 @@
 import BN from 'bn.js';
+import bs58 from 'bs58';
 
 /**
  * Exponent for calculating how many indivisible units are there in one NEAR. See {@link NEAR_NOMINATION}.
@@ -10,7 +11,7 @@ export const NEAR_NOMINATION_EXP = 24;
  */
 export const NEAR_NOMINATION = new BN('10', 10).pow(new BN(NEAR_NOMINATION_EXP, 10));
 
-// Pre-calculate offests used for rounding to different number of digits
+// Pre-calculate offsets used for rounding to different number of digits
 const ROUNDING_OFFSETS: BN[] = [];
 const BN10 = new BN(10);
 for (let i = 0, offset = new BN(5); i < NEAR_NOMINATION_EXP; i++, offset = offset.mul(BN10)) {
@@ -104,4 +105,29 @@ function formatWithCommas(value: string): string {
         value = value.replace(pattern, '$1,$2');
     }
     return value;
+}
+
+/**
+ * Encodes a Uint8Array or string into base58
+ * @param value Uint8Array or string representing a borsh encoded object
+ * @returns string base58 encoding of the value
+ */
+export function baseEncode(value: Uint8Array | string): string {
+    if (typeof value === 'string') {
+        const bytes = [];
+        for(let c = 0; c < value.length; c++){
+            bytes.push(value.charCodeAt(c));
+        }
+        value = new Uint8Array(bytes);
+    }
+    return bs58.encode(value);
+}
+
+/**
+ * Decodes a base58 string into a Uint8Array
+ * @param value base58 encoded string
+ * @returns Uint8Array representing the decoded value
+ */
+export function baseDecode(value: string): Uint8Array {
+    return new Uint8Array(bs58.decode(value));
 }