+++ packages/wallet-account/test/wallet_account.test.js
@@ -1,9 +1,10 @@
 const { KeyPair, PublicKey } = require('@near-js/crypto');
 const { InMemoryKeyStore } = require('@near-js/keystores');
 const { InMemorySigner } = require('@near-js/signers');
-const { actionCreators, createTransaction, SCHEMA, Transaction } = require('@near-js/transactions');
+const { baseDecode } = require('@near-js/utils');
+const { actionCreators, createTransaction, SCHEMA } = require('@near-js/transactions');
 const BN = require('bn.js');
-const { baseDecode, deserialize } = require('borsh');
+const { deserialize } = require('borsh');
 const localStorage = require('localstorage-memory');
 const url = require('url');
 
@@ -108,6 +109,9 @@ describe('fails gracefully on the server side (without window)', () => {
     it('throws explicit error when calling other methods on the instance', () => {
         const serverWalletConnection = new WalletConnection(nearFake, '');
         expect(() => serverWalletConnection.requestSignIn('signInContract', 'signInTitle', 'http://example.com/success',  'http://example.com/fail')).toThrow(/please ensure you are using WalletConnection on the browser/);
+        expect(() => serverWalletConnection.requestSignInUrl('signInContract', 'signInTitle', 'http://example.com/success',  'http://example.com/fail')).toThrow(/please ensure you are using WalletConnection on the browser/);
+        expect(() => serverWalletConnection.requestSignTransactions('signInContract', 'signInTitle', 'http://example.com/success',  'http://example.com/fail')).toThrow(/please ensure you are using WalletConnection on the browser/);
+        expect(() => serverWalletConnection.requestSignTransactionsUrl('signInContract', 'signInTitle', 'http://example.com/success',  'http://example.com/fail')).toThrow(/please ensure you are using WalletConnection on the browser/);
     });
 
     it('can access other props on the instance', () => {
@@ -260,8 +264,7 @@ function parseTransactionsFromUrl(urlToParse, callbackUrl = 'http://example.com/
     });
     const transactions = parsedUrl.query.transactions.split(',')
         .map(txBase64 => deserialize(
-            SCHEMA,
-            Transaction,
+            SCHEMA.Transaction,
             Buffer.from(txBase64, 'base64')));
     return transactions;
 }