+++ packages/biometric-ed25519/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @near-js/biometric-ed25519
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`662cc13d`](https://github.com/near/near-api-js/commit/662cc13d7961c3bdabed3ad51b1c57958739a3e6), [`c4655576`](https://github.com/near/near-api-js/commit/c4655576bacb1d8b85030dca5b9443649621c8ee)]:
+  - @near-js/utils@0.1.0
+  - @near-js/crypto@1.2.1
+
+## 1.1.0
+
+### Minor Changes
+
+- [#1224](https://github.com/near/near-api-js/pull/1224) [`1900c490`](https://github.com/near/near-api-js/commit/1900c49060c3ea8279448cead7347049a23f421f) Thanks [@gtsonevv](https://github.com/gtsonevv)! - remove buggy dependencies - replace js-sha256 with noble-hashes - replace elliptic with noble-curves - remove error-polyfill package
+
+### Patch Changes
+
+- [#1259](https://github.com/near/near-api-js/pull/1259) [`94cd0e65`](https://github.com/near/near-api-js/commit/94cd0e6511750beba87b823a1b87bf89f51ad7eb) Thanks [@gtsonevv](https://github.com/gtsonevv)! - pass the correct parameter to getPublicKey() in getKeys()
+
+- Updated dependencies [[`1900c490`](https://github.com/near/near-api-js/commit/1900c49060c3ea8279448cead7347049a23f421f), [`c6cdc8c7`](https://github.com/near/near-api-js/commit/c6cdc8c724a6dd53114cc5f53fd58e57cea86b78)]:
+  - @near-js/crypto@1.2.0
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`aeeb1502`](https://github.com/near/near-api-js/commit/aeeb15022a1c1deb99114eba0473739b0998fc50)]:
+  - @near-js/crypto@1.1.0
+
+## 1.0.0
+
+### Major Changes
+
+- [#1168](https://github.com/near/near-api-js/pull/1168) [`61349aec`](https://github.com/near/near-api-js/commit/61349aeca3af830f702b24654e0f13cd428192d8) Thanks [@gagdiez](https://github.com/gagdiez)! - feat: updated borsh-js to v1.0.1
+
+### Minor Changes
+
+- [#1191](https://github.com/near/near-api-js/pull/1191) [`6a67ce35`](https://github.com/near/near-api-js/commit/6a67ce35a688c6cfa4105adf0e3b06fff19f0892) Thanks [@hcho112](https://github.com/hcho112)! - Handing null response on create key flow
+
+### Patch Changes
+
+- Updated dependencies [[`695220e7`](https://github.com/near/near-api-js/commit/695220e75bc43834a7700cfc5491a7eebd324947), [`ecf29e2d`](https://github.com/near/near-api-js/commit/ecf29e2d56611a7773c79d5bb5bd20c8b7e738ea), [`61349aec`](https://github.com/near/near-api-js/commit/61349aeca3af830f702b24654e0f13cd428192d8), [`cdd8d1c8`](https://github.com/near/near-api-js/commit/cdd8d1c8c37db641bd995b2c470ad0b4fdddb93f)]:
+  - @near-js/utils@0.0.5
+  - @near-js/crypto@1.0.0
+
+## 0.3.0
+
+### Minor Changes
+
+- [#1148](https://github.com/near/near-api-js/pull/1148) [`60b59044`](https://github.com/near/near-api-js/commit/60b5904439fff25fab0cebed0b9b041938890964) Thanks [@hcho112](https://github.com/hcho112)! - bugfix and improvements on biometric-ed25519
+
+## 0.2.0
+
+### Minor Changes
+
+- [#1139](https://github.com/near/near-api-js/pull/1139) [`60c65dd1`](https://github.com/near/near-api-js/commit/60c65dd1b8609551c270a5d53b9f70970aa1bcd8) Thanks [@hcho112](https://github.com/hcho112)! - Introducing a new util function, isPassKeyAvailable on biometric-ed25519 package
+
+### Patch Changes
+
+- Updated dependencies [[`40fa6465`](https://github.com/near/near-api-js/commit/40fa64654fdaf3b463122c35521a6f72282974f2)]:
+  - @near-js/crypto@0.0.5
+
 ## 0.1.0
 
 ### Minor Changes